Restructures CSS and creates basic palette loop
