+++ src/_base.scss
@@ -21,7 +21,7 @@
 #palette {
     display: inline;
 }
-.paint {
+.paintPot {
     border-radius: 50%;
     display: inline-block;
     height: 50px;
