+++ src/_colors.scss
@@ -1,0 +1,18 @@
+@import 'variables';
+
+/* Palette Colors */
+.redPaint {
+    background-color: $red;
+}
+.greenPaint {
+    background-color: $green;
+}
+.bluePaint {
+    background-color: $blue;
+}
+.blackPaint {
+    background-color: $black;
+}
+.whitePaint {
+    background-color: $white;
+}
