+++ src/_variables.scss
@@ -1,5 +1,5 @@
 $red: #FF0000;
-$green:	#00CC00;
+$green: #00CC00;
 $blue: #0000FF;
-$white:	#FFFFFF;
+$white: #FFFFFF;
 $black: #000000;
