+++ src/components/matrix.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Cell from './cell.jsx';
-import styles from '.././index.scss';
+import styles from '.././main.scss';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import {matrixID} from './data.jsx';
 
@@ -15,6 +15,8 @@ export default class Matrix extends React.Component {
       }
       matrix.push(<div key={i} className={styles.matrixColumn}>{column}</div>)
     }
-    return(<div className={styles.matrix}>{matrix}</div>)
+    return(
+      <div className={styles.matrix}>{matrix}</div>
+    )
   }
 }
